todo: socket connect failed
