+++ lua/fzfx/env.lua
@@ -1,5 +1,4 @@
 local log = require("fzfx.log")
-local helpers = require("fzfx.helpers")
 
 local function debug_enable()
     return tostring(vim.env._FZFX_NVIM_DEBUG_ENABLE):lower() == "1"
@@ -38,23 +37,6 @@ local function search_module_path(plugin, path)
 end
 
 --- @param options Config
-local function export_fzf_default_opts(options)
-    local fzf_opts = helpers.make_fzf_opts(options)
-    if type(fzf_opts) == "string" and string.len(fzf_opts) > 0 then
-        vim.env.FZF_DEFAULT_OPTS = fzf_opts
-        log.debug(
-            "|fzfx.env - export_fzf_default_opts| FZF_DEFAULT_OPTS:%s",
-            fzf_opts
-        )
-    else
-        log.debug(
-            "|fzfx.env - export_fzf_default_opts| nil fzf_opts:%s",
-            vim.inspect(fzf_opts)
-        )
-    end
-end
-
---- @param options Config
 local function setup(options)
     -- debug
     vim.env._FZFX_NVIM_DEBUG_ENABLE = options.debug.enable and 1 or 0
@@ -73,9 +55,6 @@ local function setup(options)
             vim.env._FZFX_NVIM_ICON_ENABLE = 1
         end
     end
-
-    -- FZF_DEFAULT_OPTS
-    export_fzf_default_opts(options.fzf_opts)
 end
 
 local M = {
